Add Pipeline Parallelism to T5 & Llava

See merge request ADLR/megatron-lm!1448

+++ tests/unit_tests/dist_checkpointing/models/test_bert_model.py
@@ -28,7 +28,7 @@ def initialize_bert_model(seed, layer_spec_fn=bert_layer_with_transformer_engine
 
     layer_spec = layer_spec_fn() if callable(layer_spec_fn) else layer_spec_fn
 
-    default_config_kwargs=dict(num_layers=8, hidden_size=16, num_attention_heads=8, use_cpu_initialization=True)
+    default_config_kwargs=dict(num_layers=8, hidden_size=16, num_attention_heads=8, use_cpu_initialization=True, pipeline_dtype=torch.bfloat16)
     default_config_kwargs.update(**config_kwargs)
     transformer_config = TransformerConfig(**default_config_kwargs)
     pre_process = ps.is_pipeline_first_stage()
